# awesome_language/Go/demo/redoc/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "这是一个使用 Gin 和 Swaggo 生成的 Go 接口文档示例，并使用 ReDoc 展示。",
        "title": "Go 嵌套 API 文档示例 (ReDoc 集成)",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/users/{id}": {
            "get": {
                "description": "根据用户 ID 获取用户详细信息，包含复杂嵌套结构。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "获取用户详细信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应，包含嵌套数据",
                        "schema": {
                            "$ref": "#/definitions/main.UserResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Detail": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "字段ID",
                    "type": "integer",
                    "example": 100
                },
                "name": {
                    "description": "字段名称",
                    "type": "string",
                    "example": "Nested Item"
                }
            }
        },
        "main.UserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "嵌套的详细数据，此处即是您提到的复杂嵌套类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/main.Detail"
                        }
                    ]
                },
                "message": {
                    "description": "消息提示",
                    "type": "string",
                    "example": "Success"
                }
            }
        }
    }
}